+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- CLI arguments `--headless` and `--no-headless` to control Pianoteq GUI mode without editing config file
+
 ## [2.3.0] - 2025-11-21
 
 ### Added